Move `getTime()` into its own file.
Also make `TimeValue.Now` a bool function instead of a public field, based on PR feedback.
